OpenAI timeout (#402)
* WiP

* Add timeout for OpenAI

* Add timeout for OpenAI - float and README

* Add timeout for OpenAI - fix test

* Add timeout for OpenAI - fix test 2

* Add timeout for OpenAI - fix test 3

+++ src/OllamaConfig.php
@@ -14,7 +14,7 @@ class OllamaConfig
 
     public bool $formatJson = false;
 
-    public ?int $timeout = null;
+    public ?float $timeout = null;
 
     public ?string $apiKey = null;
